refactor: split consumer from command for agg history
